move pre-STS validation logic to non-trace property tests
